consolidate tginfo
