+++ devtools/tests/test_tginfo_models.py
@@ -13,7 +13,7 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from devtools.gearbox.tginfo_summary import collect_project_models
+from devtools.gearbox.tginfo import collect_project_models
 
 
 class TgInfoModelsTests(unittest.TestCase):
