+++ devtools/tests/test_tginfo_templates.py
@@ -12,7 +12,7 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from devtools.gearbox.tginfo_summary import collect_project_templates, format_project_templates
+from devtools.gearbox.tginfo import collect_project_templates, format_project_templates
 
 
 class FakeDecoration:
